wzg update code
Conflicts:

	source/AndroidManifest.xml
	source/src/com/V2/jni/ImRequest.java
	source/src/com/V2/jni/VideoRequest.java
	source/src/com/v2tech/service/GlobalHolder.java
	source/src/com/v2tech/view/ContactsTabFragment.java
	source/src/com/v2tech/view/ConversationsTabFragment.java
	source/src/com/v2tech/view/JNIService.java
	source/src/com/v2tech/view/conference/ConferenceCreateActivity.java
	source/src/com/v2tech/view/conference/GroupLayout.java
	source/src/com/v2tech/view/group/CrowdCreateActivity.java
	source/src/com/v2tech/vo/ConferenceGroup.java

+++ source/src/com/v2tech/service/ContactsService.java
@@ -82,7 +82,7 @@ public class ContactsService extends AbstractHandler {
 
 		mWatingGid = group.getmGId();
 		List<Group> list = GlobalHolder.getInstance().getGroup(
-				Group.GroupType.CONTACT);
+				Group.GroupType.CONTACT.intValue());
 		// Update all users which belongs this group to root group
 		if (list.size() > 0) {
 			ContactGroup defaultGroup = (ContactGroup) list.get(0);
@@ -209,7 +209,7 @@ public class ContactsService extends AbstractHandler {
 		public void onAddGroupInfo(V2Group group) {
 			if (group.type == V2Group.TYPE_CONTACTS_GROUP) {
 				Group g = new ContactGroup(group.id, group.name);
-				GlobalHolder.getInstance().addGroupToList(g.getGroupType(), g);
+				GlobalHolder.getInstance().addGroupToList(g.getGroupType().intValue(), g);
 				JNIResponse jniRes = new GroupServiceJNIResponse(
 						GroupServiceJNIResponse.Result.SUCCESS, g);
 				Message.obtain(mCallbackHandler, CREATE_CONTACTS_GROUP, jniRes)
